Fixed C data types causing an overflow error (probably since Python 3) -Python int is equivalent to C long long -Python float is equivalent to C double

+++ src/pySciCam/photron_mraw.pyx
@@ -42,9 +42,9 @@ ctypedef np.uint16_t DTYPE_t
 @cython.wraparound(False)
 @cython.nonecheck(False)
 def read_mraw(filename, int width, int height, int rgbmode = 0, tuple frames=None,\
-                          int bits_per_pixel=12, long start_offset = 0, int quiet = 0):
+                          int bits_per_pixel=12, long long start_offset = 0, int quiet = 0):
 
-    cdef long t0 = time.time()
+    cdef double t0 = time.time()
     cdef double bytes_per_pixel
     if rgbmode == 1: bytes_per_pixel = 3.0*bits_per_pixel/8.0
     else: bytes_per_pixel = bits_per_pixel/8.0
@@ -55,7 +55,7 @@ def read_mraw(filename, int width, int height, int rgbmode = 0, tuple frames=Non
 
     # Get size of binary file before we start reading
     # (it might be smaller than we think)
-    cdef long nbytes = os.path.getsize(filename)
+    cdef long long nbytes = os.path.getsize(filename)
 
     # Scanlines are padded to nearest 16 bytes
     cdef unsigned int scanline_pad = int((width*1.5)%16)
@@ -90,8 +90,8 @@ def read_mraw(filename, int width, int height, int rgbmode = 0, tuple frames=Non
 
 
     # Limit number of frames loaded from file (good for testing)
-    cdef long start = start_offset # bytes
-    cdef long end = nbytes
+    cdef long long start = start_offset # bytes
+    cdef long long end = nbytes
     if frames is not None:
         if frames[1] > 0:
             start = start_offset + int(bytes_per_frame*frames[0])
